Reverted to sample project
Forget what I said in the previous commit. Doing that still left the
issue that R was unable to be defined. I imported the sample project
that would be at the same place in the tutorial to see if it worked.

The only thing I can think of to explain the issues would be differences
in standards and versions. Considering that the tutorial I'm following
is not current and a bit outdated (in terms of Eclipse and Android) it
may have had conflicts that I don't understand.